requirements.txt
Flask==3.0.1
python-dotenv==1.0.1
python-telegram-bot[job-queue]==20.7
python_dateutil==2.8.2
